6.4.10 - Gutenberg Improvements
- Wrapping block functionality in a try catch
- Consolidating JS functions into a single class file
- Consolidating JS constants into a single constants file
- Check if no forms are created and show a message if so
- Check if API key is empty/invalid and show a message if so
- Default the reCAPTCHA to the plugin's reCAPTCHA settings
